Use Ruby from `.ruby-version` if it exists
Currently, we use the latest version of Ruby available to run this
GitHub Action. To specify a particular Ruby version, one can make use of
`with:` and set the `ruby-version:`. As a convention, Ruby projects
allow the Ruby version to be set in a file at the root of the project
called `.ruby-version`.

This commit updates this GitHub Action to read the Ruby version from
`.ruby-version` if it exists. If not, it will fall back to the defaults.

Ref:
- https://github.com/standardrb/standard-ruby-action?tab=readme-ov-file#options
- https://github.com/standardrb/standard-ruby-action/issues/19