Create Release GHA workflow
This workflow contains two jobs that will each run on git-tag pushes, or
direct workflow dispatch (through the github UI).

The `github` job will create a GitHub Release for the triggering git
tag. If the tag name contains `-` (signifying prerelease per semver),
the release is marked as a prerelease. Release notes are generated
automatically using the gh-cli (handled by github). A future enhancement
would be to pull them from the Changelog.

The `vMajor-ref` job pushes a `vMajor` ref corresponding to the
triggering tag. For instance, if the latest tag is `v1.2.3`, and
`v1.2.4` or `v1.3.0` are pushed, then the branch `v1` is force-pushed to
point to the latest tag. This way these `vMajor` branches can be used by
users to pin to a "semver major" version rather than a minor or patch
release. (In actuality, this practice should probably be retired in
favor of using dependabot to bump and pin actions to explicit git-shas.)

This fixes #26 and #8