Pin actions to git shas for security
The v* refs (either tag or branch), are mutable. (They can be
force-pushed.) Which means, they can change _under your feet_.

It is marginally more secure to pin to specific shas, instead of
arbitrary (mutable) refs.
(see "pin actions to full commit sha":
https://docs.github.com/en/actions/how-tos/security-for-github-actions/security-guides/security-hardening-for-github-actions#using-third-party-actions )

Adds a trailing comment to the sha which indicates the corresponding
version tag. This works because dependabot itself not only bumps the
sha, but will bump the trailing comment as well, so we don't lose the
human-consumable context of what released version is being used.

https://github.blog/changelog/2022-10-31-dependabot-now-updates-comments-in-github-actions-workflows-referencing-action-versions/