Naïvely check for ruby specifier in asdf/mise files
Checking only for existence of .tool-versions or mise.toml files is ripe
for false-positives, as those files could exist for specifying only node
versions (or a bunch of other tools) and may not define a ruby version
at all. This would cause setup-ruby to fail because we would omit an
explicit ruby-version input to setup-ruby thinking it would find in
those files, but it would not. And since setup-ruby does not have a
final fallback, it would error.

Now we check for existence of `ruby` string in those files, as must
exist for those files to apply to ruby. Technically, there could still
be false positives, (the `ruby` string could be in a comment, or those
files could be malformed per their parsing rules). But those scenarios
would be extremely low. This quick check should resolve virtually all
false positives. (Any remaining false positives, should compel the user
to explicitly configure a ruby-version parameter as an input.)