Rename workdir to working-directory
working-directory is the commonly used name in GitHub Actions own
syntax, and is also the name used by setup-ruby action. We should
conform to that name, instead of having a name that means the same
thing, but spelled differently.

The old input now defaults to empty string, which is falsy. So we use it
if truthy (present), otherwise we fallback to the new input param which
has the expected default value. A deprecation message is printed to give
a warning to users to rename the input.